Move event/ folder to input/
